Add util function to check migration status
